更新 tools/get_all_package_url.sh 脚本
增加支持对部分域名使用跳过证书认证下载方式。

+++ loongarch32/tools/get_all_package_url.sh
@@ -41,6 +41,8 @@ declare SINGLE_PACKAGE=FALSE
 declare VERSION_INDEX=""
 declare WORLD_PARM=""
 declare AUTO_DOWNLOAD_OR_COPY=0
+declare SAVE_PROXY=""
+declare EMU_BROWSER=""
 
 while getopts 'gfuri:pswav:h' OPT; do
     case $OPT in
@@ -58,6 +60,7 @@ while getopts 'gfuri:pswav:h' OPT; do
 	    ;;
 	p)
 	    PROXY_MODE=1
+	    SAVE_PROXY=${https_proxy}
 	    ;;
 	i)
 	    INDEX_FILE=$OPTARG
@@ -114,17 +117,63 @@ function set_proxy
         declare DOMAIN=$(echo ${1} | grep -o ".\{0,\}//[^/]\{0,\}" | awk -F'/' '{print $NF}')
         declare PROXY_SERVER=""
 	if [ -f proxy.set ]; then 
-		PROXY_SERVER=$(cat proxy.set | grep -v "^#" | grep "${DOMAIN} " | awk -F' ' '{ print $2 }')
+# 		PROXY_SERVER=$(cat proxy.set | grep -v "^#" | grep "${DOMAIN} " | awk -F' ' '{ print $2 }')
+        	PROXY_SERVER=$(cat proxy.set | grep -v "^#" | grep "${DOMAIN} " | awk -F' ' '{ print $2 }' | awk -F'|' '{ print $1 }')
+        	PROXY_RETRY=$(cat proxy.set | grep -v "^#" | grep "${DOMAIN} " | awk -F' ' '{ print $2 }' | awk -F'|' '{ print $2 }')
+        	PROXY_EMU=$(cat proxy.set | grep -v "^#" | grep "${DOMAIN} " | awk -F' ' '{ print $2 }' | awk -F'|' '{ print $3 }')
 	fi
-        if [ "x${PROXY_SERVER}" == "x" ]; then
-                        unset https_proxy
-                        unset http_proxy
+	if [ "x${PROXY_EMU}" == "x" ]; then
+		EMU_BROWSER=""
+	else
+		EMU_BROWSER="${PROXY_EMU}"
+	fi
+#         if [ "x${PROXY_SERVER}" == "x" ]; then
+#                         unset https_proxy
+#                         unset http_proxy
+#         else
+#                         export https_proxy=${PROXY_SERVER}
+#                         export http_proxy=${PROXY_SERVER}
+#         fi
+
+	if [ "x${PROXY_SERVER}" == "x" ]; then
+		if [ "x${SAVE_PROXY}" == "x" ]; then
+#                       echo "取消代理"
+			unset https_proxy
+                else
+# 			echo "恢复代理: ${SAVE_PROXY}"
+			export https_proxy=${SAVE_PROXY}
+                fi
         else
-                        export https_proxy=${PROXY_SERVER}
-                        export http_proxy=${PROXY_SERVER}
+# 		echo "设置代理：https_proxy=${PROXY_SERVER}"
+		export https_proxy=${PROXY_SERVER}
         fi
 }
 
+function replace_url
+{
+	declare DOMAIN=$(echo ${1} | grep -o ".\{0,\}//[^/]\{0,\}" | awk -F'/' '{print $NF}')
+	declare REPLACE_URL=""
+	if [ -f replace.url ]; then
+		REPLACE_URL=$(cat replace.url | grep -v "^#" | grep "${DOMAIN} " | awk -F' ' '{ print $2 }')
+		if [ "x${REPLACE_URL}" == "x" ]; then
+			REPLACE_URL=$(cat replace.url | grep -v "^#" | grep "^* " | awk -F' ' '{ print $2 }')
+		fi
+	fi
+
+	if [ "x${REPLACE_URL}" == "x" ]; then
+		echo "${1}"
+	else
+		case "${REPLACE_URL%%/*}" in
+			https: | http: | ftps: | ftp:)
+				echo "${REPLACE_URL}${2}"
+				;;
+			*)
+				echo "${1}"
+				;;
+		esac
+			
+	fi
+}
 
 function get_pkg_old_version
 {
@@ -141,6 +190,13 @@ function get_pkg_old_version
 		return ""
 	fi
 }
+
+
+# function test_git_need_clone
+# {
+# 	echo "0"
+# 	return
+# }
 
 
 mkdir -p ${NEW_BASE_DIR}/downloads/sources/{files,hash}
@@ -196,6 +252,7 @@ do
 		URL="$(replace_arch_parm "$(cat ${SOURCES_DIR}/url/${i}.${VERSION_INDEX} | awk -F'|' '{ print $2 }')" )"
 		SAVE_FILENAME="$(replace_arch_parm "$(cat ${SOURCES_DIR}/url/${i}.${VERSION_INDEX} | awk -F'|' '{ print $3 }')" )"
 	fi
+	EMU_BROWSER=""
 
 	if [ "x${SAVE_FILENAME}" == "x" ]; then
 		SAVE_FILENAME="${URL##*/}"
@@ -229,12 +286,34 @@ do
 						fi
 						echo "下载：$i 所需源码包到${NEW_BASE_DIR}/downloads/sources/files/${SAVE_FILENAME}..."
 						# wget -c ${URL} -O ${BASE_DIR}/downloads/sources/files/${URL##*/}
-						wget -c ${URL} -O ${NEW_BASE_DIR}/downloads/sources/files/${SAVE_FILENAME}
+						REPLACE_REAL_URL=$(replace_url "${URL}" "${SAVE_FILENAME}")
+# 						echo "EMU_BROWSER = ${EMU_BROWSER}"
+						if [ "x${EMU_BROWSER}" == "x" ]; then
+							wget -c ${REPLACE_REAL_URL} -O ${NEW_BASE_DIR}/downloads/sources/files/${SAVE_FILENAME}
+						else
+							wget -c ${REPLACE_REAL_URL} -O ${NEW_BASE_DIR}/downloads/sources/files/${SAVE_FILENAME} --no-check-certificate
+						fi
 						if [ "x$?" != "x0" ]; then
-							echo "${URL} 下载失败！"
-							echo "下载 ${URL} 失败！" >> logs/download_fail.log
-							((FAIL_COUNT++))
-							continue;
+							if [ "${REPLACE_REAL_URL}" == "${URL}" ]; then
+								echo "${URL} 下载失败！"
+								echo "下载 ${URL} 失败！" >> logs/download_fail.log
+								((FAIL_COUNT++))
+								continue;
+							else
+								if [ "x${EMU_BROWSER}" == "x" ]; then
+									wget -c ${URL} -O ${NEW_BASE_DIR}/downloads/sources/files/${SAVE_FILENAME}
+								else
+									wget -c ${URL} -O ${NEW_BASE_DIR}/downloads/sources/files/${SAVE_FILENAME} --no-check-certificate
+								fi
+								if [ "x$?" != "x0" ]; then
+									echo "${URL} 与 ${REPLACE_REAL_URL} 均下载失败！"
+									echo "下载 ${URL} ( 提换下载地址：${REPLACE_REAL_URL} ) 均下载失败！" >> logs/download_fail.log
+									((FAIL_COUNT++))
+									continue;
+								fi
+							fi
+# 							((FAIL_COUNT++))
+# 							continue;
 						fi
 						md5sum ${NEW_BASE_DIR}/downloads/sources/files/${SAVE_FILENAME} | awk -F' ' '{print $1}' > ${NEW_BASE_DIR}/downloads/sources/hash/${SAVE_FILENAME}.hash
 					fi
@@ -394,7 +473,7 @@ ${i} 没有下载路径，请检查。"
 	fi
 
 
-	if [ -d ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ ] && [ "x$(find ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ -name *.url)" != "x" ]; then
+	if [ -d ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ ] && [ "x$(find ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ -name "*.url")" != "x" ]; then
 		echo "发现${i}存在需要下载的资源文件……"
 		for url_i in $(ls ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/*.url)
 		do
@@ -471,12 +550,26 @@ ${i} 没有下载路径，请检查。"
 									if [ -f ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${RESOURCES_FILENAME} ]; then
 										rm -f ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${RESOURCES_FILENAME}
 									fi
-									wget -c ${RESOURCES_URL} -O ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${RESOURCES_FILENAME}
+# 									wget -c ${RESOURCES_URL} -O ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${RESOURCES_FILENAME}
+									REPLACE_REAL_RESOURCES_URL=$(replace_url "${RESOURCES_URL}" "${RESOURCES_FILENAME}")
+									wget -c ${REPLACE_REAL_RESOURCES_URL} -O ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${RESOURCES_FILENAME}
 									if [ "x$?" != "x0" ]; then
-										echo "${RESOURCES_URL} 下载失败！"
-										echo "下载${i}资源文件 ${RESOURCES_URL} 失败！" >> logs/download_fail.log
-										((FAIL_COUNT++))
-										continue;
+										if [ "${REPLACE_REAL_RESOURCES_URL}" == "${RESOURCES_URL}" ]; then
+											echo "${RESOURCES_URL} 下载失败！"
+											echo "下载 ${i} 的资源文件 ${RESOURCES_URL} 失败！" >> logs/download_fail.log
+											((FAIL_COUNT++))
+											continue;
+										else
+											wget -c ${RESOURCES_URL} -O ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${RESOURCES_FILENAME}
+											if [ "x$?" != "x0" ]; then
+												echo "${RESOURCES_URL} 与 ${REPLACE_REAL_RESOURCES_URL} 均下载失败！"
+												echo "下载 ${i} 的资源文件 ${RESOURCES_URL} ( 提换下载地址：${REPLACE_REAL_RESOURCES_URL} ) 均下载失败！" >> logs/download_fail.log
+												((FAIL_COUNT++))
+												continue;
+											fi
+										fi
+# 										((FAIL_COUNT++))
+# 										continue;
 									fi
 									md5sum ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${RESOURCES_FILENAME} | awk -F' ' '{print $1}' > ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/hash/${RESOURCES_FILENAME}.hash
 								fi
@@ -495,7 +588,7 @@ ${i} 没有下载路径，请检查。"
 		echo "已完成${i}需要下载的资源文件。"
 	fi
 
-	if [ -d ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ ] && [ "x$(find ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ -name *.listfile)" != "x" ]; then
+	if [ -d ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ ] && [ "x$(find ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/ -name "*.listfile")" != "x" ]; then
 		if [ "x${GIT_ONLY}" == "xFALSE" ]; then
 			echo "发现${i}存在需要下载的资源组文件……"
 			for listfile_i in $(ls ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/*.listfile)
@@ -538,12 +631,26 @@ ${i} 没有下载路径，请检查。"
 								echo "从${LIST_URL}下载${LIST_FILENAME}中的文件..."
 								mkdir -p ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/{files,hash}/
 								mkdir -p ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${LIST_NAME}_dir
-								wget -c -B ${LIST_URL} -i ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/${LIST_FILENAME} -P ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${LIST_NAME}_dir/
+# 								wget -c -B ${LIST_URL} -i ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/${LIST_FILENAME} -P ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${LIST_NAME}_dir/
+								REPLACE_REAL_LIST_URL=$(replace_url "${LIST_URL}" "")
+								wget -c -B ${REPLACE_REAL_LIST_URL} -i ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/${LIST_FILENAME} -P ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${LIST_NAME}_dir/
 								if [ "x$?" != "x0" ]; then
-									echo "从 ${LIST_URL} 下载资源组文件 ${LIST_FILENAME} 失败！"
-									echo "${i}从${LIST_URL}下载资源组文件 ${LIST_FILENAME} 失败！" >> logs/download_fail.log
-									((FAIL_COUNT++))
-									continue;
+									if [ "${REPLACE_REAL_LIST_URL}" == "${LIST_URL}" ]; then
+										echo "从 ${LIST_URL} 下载资源组文件 ${LIST_FILENAME} 失败！"
+										echo "${i} 从 ${LIST_URL} 下载资源组文件 ${LIST_FILENAME} 失败！" >> logs/download_fail.log
+										((FAIL_COUNT++))
+										continue;
+									else
+										wget -c -B ${LIST_URL} -i ${NEW_BASE_DIR}/files/step/${i}/${PKG_VERSION}/${LIST_FILENAME} -P ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${LIST_NAME}_dir/
+										if [ "x$?" != "x0" ]; then
+											echo "从 ${LIST_URL} 与 ${REPLACE_REAL_LIST_URL} 下载资源组文件 ${LIST_FILENAME} 均失败！"
+											echo "${i} 从 ${LIST_URL} ( 提换下载地址：${REPLACE_REAL_LIST_URL} ) 下载资源组文件 ${LIST_FILENAME} 均失败！" >> logs/download_fail.log
+											((FAIL_COUNT++))
+											continue;
+										fi
+									fi
+# 									((FAIL_COUNT++))
+# 									continue;
 								fi
 								pushd ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/ > /dev/null
 									tar -czf ${NEW_BASE_DIR}/downloads/files/step/${i}/${PKG_VERSION}/files/${LIST_NAME}_dir.tar.gz ${LIST_NAME}_dir
